+++ QuICT/core/gate/composite_gate.py
@@ -342,7 +342,7 @@ class CompositeGate(list):
                 else:
                     max_qubit = max(max_qubit, arg)
         if min_qubit == -1:
-            return np.eye(2, dtype=np.complex128)
+            return np.eye(2, dtype=np.complex64)
 
         if local:
             q_len = max_qubit - min_qubit + 1
@@ -351,10 +351,10 @@ class CompositeGate(list):
 
         n = 1 << q_len
 
-        result = np.eye(n, dtype=np.complex128)
+        result = np.eye(n, dtype=np.complex64)
 
         for gate in self.gates:
-            new_values = np.zeros((n, n), dtype=np.complex128)
+            new_values = np.zeros((n, n), dtype=np.complex64)
             targs = gate.affectArgs
             for i in range(len(targs)):
                 targs[i] -= min_qubit
